added checking for cookies to each appropriate page

+++ server/handlers/deleted.js
@@ -4,13 +4,18 @@ var sqlite = require("sqlite3");
 var db = new sqlite.Database("albums.db");
 
 module.exports = function(req, reply){
+  if(!req.state.username){
+    reply.redirect("/login");
+  } else {
     fs.readFile("models/id.json", "utf8", function(err, data){
       console.log(data);
       db.all("DELETE FROM tracks WHERE albumID='"+data+"';");
       db.all("DELETE FROM albums WHERE albumId='"+data+"';");
       reply.view("deleted", {
-        message: "Deletion successful!"
+        message: "Deletion successful!",
+        user: req.state.username
       });
       console.log("Deleted!");
     });
+  }
 };
